Hide autostyling button for collapsed panel.
